Convert List of Samples into a Data Frame

This (internal) function takes a list with differ length data.frames or vectors and pads them all to make a clean data.frame.

Usage

.convertSamplesToDF(data_list)

Arguments

data_list

List of elements to be combined to a data.frame.

Value

Data.frame of the data in data_list

Details

This is an internal function and will not be viewable to user. See generalized_resampling for usage