Games-Howell Test

Computes Games-Howell test on data. Similar to Tukey HSD, it computes the significance of the differences between groups. However, it requires far fewer assumptions. Per the original paper, non-normality it not a problems, groups can be different sizes, and there is assumption on the homogeneity of variances. Further, small sample sizes are okay (each group is typically recommended to have at least 6 observations).

Usage

games_howell(data, alpha = 0.05)

Arguments

data

Data.frame with the first column the values and the second column the group names

alpha

Significance for confidence intervals, defaults to 0.05

Value

Table showing groups, their differences, and significance, among other details

References

Games, P. A., & Howell, J. F. (1976). Pairwise Multiple Comparison Procedures with Unequal N’s and/or Variances: A Monte Carlo Study. Journal of Educational Statistics, 1(2), 113–125.

Games, P. A., Keselman, H. J., & Clinch, J. J. (1979). Tests for homogeneity of variance in factorial designs. Psychological Bulletin, 86(5), 978–984.

See also

Examples

data <- data.frame('value'=c(rnorm(14,sd = 2), rnorm(6), rnorm(20,mean=2)),
                   'group'=c(rep('A',14),rep('B',6), rep('C',20)))
games_howell(data)
#>         diff         lwr      upr        se        t       df           p
#> B-A 0.818905 -0.98497033 2.622780 0.4948727 1.170105 16.17951 0.486914365
#> C-A 2.398931  0.91089704 3.886965 0.4137928 4.099396 18.78851 0.001719933
#> C-B 1.580026  0.09833288 3.061719 0.3687404 3.029901  8.22190 0.037766316