Subset of the Anatomical Data from Domestic Cats

A subset of cats. The heart and body weights of samples of male and female cats used for digitalis experiments. The cats were all adult, over 2 kg body weight.

Usage

subcats

Format

subcats

A data frame with 116 rows and 3 columns:

Sex

Sex of cat as a factor with levels F and M

Bwt

Body weight in kilograms

Hwt

Heart weight in kilograms

Source

Fisher, R.A. (1947). The analysis of covariance method for the relation between a part and the whole. Biometrics, 3 2, 65-8.

References

Venables, W. N. & Ripley, B. D. (2002) Modern Applied Statistics with S. Fourth Edition. Springer, New York. ISBN 0-387-95457-0